# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"TripDifficulty": {
				"enum": [
					"easy",
					"moderate",
					"challenging",
					"strenuous",
					"extreme"
				],
				"type": "string"
			},
			"TrekStatus": {
				"enum": [
					"active",
					"featured",
					"draft"
				],
				"type": "string"
			},
...
			"ApiResponse_Trek-Array_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/Trek"
								},
								"type": "array"
							},
							{
								"items": {
									"items": {
										"$ref": "#/components/schemas/Trek"
									},
									"type": "array"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_Trek_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Trek"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Trek"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_boolean_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"type": "boolean"
							},
							{
								"items": {
									"type": "boolean"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TourType": {
				"enum": [
					"cultural_heritage",
					"luxury_wellness",
					"wildlife_safari",
					"helicopter_tour",
					"day_tour",
					"other"
				],
				"type": "string"
			},
			"TourStatus": {
				"enum": [
					"active",
					"featured",
					"draft"
				],
				"type": "string"
			},
...
			"ApiResponse_Tour-Array_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/Tour"
								},
								"type": "array"
							},
							{
								"items": {
									"items": {
										"$ref": "#/components/schemas/Tour"
									},
									"type": "array"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Tour_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Tour"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Tour"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ClimbingGrade": {
				"enum": [
					"Non-Technical Trekking Peak",
					"Technical Alpine Grade",
					"Extreme Technical Grade"
				],
				"type": "string"
			},
			"ExpeditionStatus": {
				"enum": [
					"active",
					"featured",
					"draft"
				],
				"type": "string"
			},
...
			"ApiResponse_Expedition-Array_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/Expedition"
								},
								"type": "array"
							},
							{
								"items": {
									"items": {
										"$ref": "#/components/schemas/Expedition"
									},
									"type": "array"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Expedition_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Expedition"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Expedition"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"BlogStatus": {
				"enum": [
					"published",
					"draft",
					"archived"
				],
				"type": "string"
			},
...
			"ApiResponse_BlogArticle-Array_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/BlogArticle"
								},
								"type": "array"
							},
							{
								"items": {
									"items": {
										"$ref": "#/components/schemas/BlogArticle"
									},
									"type": "array"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_BlogArticle_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/BlogArticle"
							},
							{
								"items": {
									"$ref": "#/components/schemas/BlogArticle"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "rest-nodejs-starter",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/treks/admin": {
			"get": {
				"operationId": "GetAdminAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Trek-Array_"
								}
							}
						}
					}
				},
				"description": "Admin-only listing of ALL trekking packages (including DRAFT, ACTIVE, and FEATURED).",
				"tags": [
					"Trekking"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "categoryId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "region",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "difficulty",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TripDifficulty"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TrekStatus"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "minPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "minDuration",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxDuration",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "minAltitude",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxAltitude",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/treks/{idOrSlug}": {
			"get": {
				"operationId": "GetByIdOrSlug",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Trek_"
								}
							}
						}
					}
				},
				"description": "Get trekking package by ID or Slug (public active/featured only).",
				"tags": [
					"Trekking"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "idOrSlug",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/tours/admin": {
			"get": {
				"operationId": "GetAdminAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Tour-Array_"
								}
							}
						}
					}
				},
				"description": "Admin-only listing of ALL tour packages (including DRAFT, ACTIVE, and FEATURED).",
				"tags": [
					"Tours"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "categoryId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "region",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tourType",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TourType"
						}
					},
					{
						"in": "query",
						"name": "difficulty",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TripDifficulty"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TourStatus"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "minPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "minDuration",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxDuration",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/tours/{idOrSlug}": {
			"get": {
				"operationId": "GetByIdOrSlug",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Tour_"
								}
							}
						}
					}
				},
				"description": "Get tour package by ID or Slug (public active/featured only).",
				"tags": [
					"Tours"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "idOrSlug",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/expeditions": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Expedition-Array_"
								}
							}
						}
					}
				},
				"description": "Get all active expedition packages with dynamic filtering.",
				"tags": [
					"Expeditions"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "categoryId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "region",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "difficulty",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TripDifficulty"
						}
					},
					{
						"in": "query",
						"name": "climbingGrade",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ClimbingGrade"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ExpeditionStatus"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "minPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "minAltitude",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxAltitude",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "minPeakHeight",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxPeakHeight",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Expedition_"
								}
							}
						}
					}
				},
				"description": "Create a new expedition package.",
				"tags": [
					"Expeditions"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateExpeditionDto"
							}
						}
					}
				}
			}
		},
		"/expeditions/admin": {
			"get": {
				"operationId": "GetAdminAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Expedition-Array_"
								}
							}
						}
					}
				},
				"description": "Admin-only listing of ALL expedition packages (including DRAFT, ACTIVE, and FEATURED).",
				"tags": [
					"Expeditions"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "categoryId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "region",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "difficulty",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TripDifficulty"
						}
					},
					{
						"in": "query",
						"name": "climbingGrade",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ClimbingGrade"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ExpeditionStatus"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "minPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "minAltitude",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxAltitude",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "minPeakHeight",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxPeakHeight",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/expeditions/{idOrSlug}": {
			"get": {
				"operationId": "GetByIdOrSlug",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Expedition_"
								}
							}
						}
					}
				},
				"description": "Get expedition package by ID or Slug (public active/featured only).",
				"tags": [
					"Expeditions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "idOrSlug",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/expeditions/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Expedition_"
								}
							}
						}
					}
				},
				"description": "Update an existing expedition package.",
				"tags": [
					"Expeditions"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateExpeditionDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_boolean_"
								}
							}
						}
					}
				},
				"description": "Delete an expedition package.",
				"tags": [
					"Expeditions"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/blogs/admin": {
			"get": {
				"operationId": "GetAdminAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_BlogArticle-Array_"
								}
							}
						}
					}
				},
				"description": "Admin-only listing of ALL blog articles (including DRAFT, PUBLISHED, and ARCHIVED).",
				"tags": [
					"Blog Articles"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/BlogStatus"
						}
					},
					{
						"in": "query",
						"name": "categoryId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/blogs/{idOrSlug}": {
			"get": {
				"operationId": "GetByIdOrSlug",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_BlogArticle_"
								}
							}
						}
					}
				},
				"tags": [
					"Blog Articles"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "idOrSlug",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/admin/treks/{idOrSlug}": {
			"get": {
				"operationId": "GetByIdOrSlug",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Trek_"
								}
							}
						}
					}
				},
				"description": "Admin retrieve trekking package by ID or Slug.",
				"tags": [
					"Admin Trekking Management"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "idOrSlug",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/admin/tours/{idOrSlug}": {
			"get": {
				"operationId": "GetByIdOrSlug",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Tour_"
								}
							}
						}
					}
				},
				"description": "Admin retrieve tour package by ID or Slug.",
				"tags": [
					"Admin Tours Management"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "idOrSlug",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/admin/expeditions": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Expedition-Array_"
								}
							}
						}
					}
				},
				"description": "Admin-only listing of ALL expedition packages (including DRAFT, ACTIVE, and FEATURED).",
				"tags": [
					"Admin Expeditions Management"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "categoryId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "region",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "difficulty",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TripDifficulty"
						}
					},
					{
						"in": "query",
						"name": "climbingGrade",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ClimbingGrade"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ExpeditionStatus"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "minPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "minAltitude",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxAltitude",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "minPeakHeight",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxPeakHeight",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Expedition_"
								}
							}
						}
					}
				},
				"description": "Create a new expedition package.",
				"tags": [
					"Admin Expeditions Management"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateExpeditionDto"
							}
						}
					}
				}
			}
		},
		"/admin/expeditions/{idOrSlug}": {
			"get": {
				"operationId": "GetByIdOrSlug",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Expedition_"
								}
							}
						}
					}
				},
				"description": "Admin retrieve expedition package by ID or Slug.",
				"tags": [
					"Admin Expeditions Management"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "idOrSlug",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/expeditions/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Expedition_"
								}
							}
						}
					}
				},
				"description": "Update an existing expedition package.",
				"tags": [
					"Admin Expeditions Management"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateExpeditionDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_boolean_"
								}
							}
						}
					}
				},
				"description": "Delete an expedition package.",
				"tags": [
					"Admin Expeditions Management"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/admin/blogs/{idOrSlug}": {
			"get": {
				"operationId": "GetByIdOrSlug",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_BlogArticle_"
								}
							}
						}
					}
				},
				"description": "Admin retrieve blog article by ID or Slug.",
				"tags": [
					"Admin Blogs Management"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "idOrSlug",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "https://abc.com/api/v1",
			"description": "prouduction server"
		},
		{
			"url": "http://localhost:4000/api/v1",
			"description": "Development server"
		}
	]
}